Force final save to use full validations so it doesn't break business logic.
This undoes the (dangerous) override of valid? and instead adds a
valid_for_validation_group method.  In this way valid? will always respect the
business logic validations added on a model as a user would expect, but one
can check for a particular validation group still.

The wizard flow then uses the new method on each step, it still uses the
save_without_validation for the persisting of the model between steps, but now
it does a normal save! on complete so if there are any programming errors that
would cause an invalid model the generate an exception instead of making their
way into the DB.